Correcting of code

+++ Labs (C++)/Lab. work 4/src/Lab. work 4.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main()
 {
-	unsigned n;
+	unsigned int n;
 	cout << "Enter a number which is equal to 4 or is bigger than the last one: ";
 	cin >> n;	// Initialization of n
 	while (n < 4)	// Checking of n for performing of the condition
@@ -15,12 +15,12 @@ int main()
 		cin >> n;
 	}
 
-	long double v3 = 0, v2 = 0, v1 = 1.5, v; // Initialization of three primary elements of the sequence
-	long double &vn = v;
+	double v3 = 0, v2 = 0, v1 = 1.5, v; // Initialization of three primary elements of the sequence
+	double &vn = v;
 
-	for (unsigned i = 4; i <= n; i++)	// Calculating of vn
+	for (unsigned int i = 4; i <= n; i++)	// Calculating of vn
 	{
-		v = (i + static_cast<unsigned>(1)) / static_cast<long double>(pow(i, 2) + 1) * v1 - v2 * v3;
+		v = static_cast<double>(i + 1.) / static_cast<double>(pow(static_cast<double>(i), 2) + 1.) * v1 - v2 * v3;
 		v3 = v2;
 		v2 = v1;
 		v1 = v;
